requirements.txt
discord-py-interactions
python-decouple
pytest